+++ unit-ui/src/widgets/input.rs
@@ -77,6 +77,10 @@ impl<'a> BasicInput<'a> {
     }
 }
 
+impl Widget for BasicInput<'_> {
+    fn render(self, area: Rect, buf: &mut Buffer) { (&self).render(area, buf); }
+}
+
 impl Widget for &BasicInput<'_> {
     fn render(self, area: Rect, buf: &mut Buffer) {
         let border = Style::default().fg(if self.focused {